form player operations
